# docs/swagger/swagger.yaml
basePath: /
...
definitions:
...
  github_com_LerianStudio_matcher_internal_matching_adapters_http_dto.MatchGroupResponse:
    description: Aggregate of matched transactions
    properties:
      confidence:
        description: Confidence score (0-100)
        example: 85
        maximum: 100
        minimum: 0
        type: integer
      confirmedAt:
        description: When the group was confirmed in RFC3339 format
        example: "2025-01-15T10:35:00Z"
        type: string
      contextId:
        description: Context ID this group belongs to
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      createdAt:
        description: Creation timestamp in RFC3339 format
        example: "2025-01-15T10:30:00Z"
        type: string
      id:
        description: Unique identifier for the group
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      items:
        description: Matched items in this group
        items:
          $ref: '#/definitions/github_com_LerianStudio_matcher_internal_matching_adapters_http_dto.MatchItemResponse'
        type: array
      rejectedReason:
        description: Reason for rejection if rejected
        example: amounts do not match within tolerance
        type: string
      ruleId:
        description: Rule ID used for matching (null for manual matches)
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      runId:
        description: Run ID that created this group
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      status:
        description: Current status of the group
        enum:
        - PROPOSED
        - CONFIRMED
        - REJECTED
        example: PROPOSED
        type: string
      updatedAt:
        description: Last update timestamp in RFC3339 format
        example: "2025-01-15T10:30:00Z"
        type: string
    type: object
  github_com_LerianStudio_matcher_internal_matching_adapters_http_dto.MatchItemResponse:
    description: Allocation of a transaction within a match group
    properties:
      allocatedAmount:
        description: Amount allocated in this match
        example: "1000.00"
        type: string
      allocatedCurrency:
        description: Currency of the allocated amount
        example: USD
        type: string
      allowPartial:
        description: Whether partial allocation is allowed
        example: false
        type: boolean
      createdAt:
        description: Creation timestamp in RFC3339 format
        example: "2025-01-15T10:30:00Z"
        type: string
      expectedAmount:
        description: Expected amount for full match
        example: "1000.00"
        type: string
      id:
        description: Unique identifier for the item
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      matchGroupId:
        description: Match group ID this item belongs to
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      transactionId:
        description: Transaction ID being matched
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      updatedAt:
        description: Last update timestamp in RFC3339 format
        example: "2025-01-15T10:30:00Z"
        type: string
    type: object
  github_com_LerianStudio_matcher_internal_matching_adapters_http_dto.MatchRunResponse:
    description: Execution of the matching engine
    properties:
      completedAt:
        description: When the run completed in RFC3339 format (if completed)
        example: "2025-01-15T10:35:00Z"
        type: string
      contextId:
        description: Context ID this run belongs to
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      createdAt:
        description: Creation timestamp in RFC3339 format
        example: "2025-01-15T10:30:00Z"
        type: string
      failureReason:
        description: Failure reason if run failed
        example: no matching rules found
        type: string
      id:
        description: Unique identifier for the run
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      mode:
        description: Execution mode
        enum:
        - DRY_RUN
        - COMMIT
        example: DRY_RUN
        type: string
      startedAt:
        description: When the run started in RFC3339 format
        example: "2025-01-15T10:30:00Z"
        type: string
      stats:
        additionalProperties:
          type: integer
        description: Run statistics (matched count, unmatched count, etc.)
        type: object
      status:
        description: Current status of the run
        enum:
        - PROCESSING
        - COMPLETED
        - FAILED
        example: COMPLETED
        type: string
      updatedAt:
        description: Last update timestamp in RFC3339 format
        example: "2025-01-15T10:30:00Z"
        type: string
    type: object
  http.ErrorResponse:
    description: Standard error response returned by all API endpoints
    properties:
      code:
        description: HTTP status code
        example: 400
        type: integer
      message:
        description: Human-readable error message
        example: context name is required
        type: string
      title:
        description: Error type identifier
        example: invalid_request
        type: string
    type: object
...
  internal_matching_adapters_http.ListMatchGroupsResponse:
    description: Cursor-paginated list of match groups
    properties:
      hasMore:
        example: true
        type: boolean
      items:
        description: List of match groups
        items:
          $ref: '#/definitions/github_com_LerianStudio_matcher_internal_matching_adapters_http_dto.MatchGroupResponse'
        maxItems: 200
        type: array
      limit:
        example: 20
        maximum: 200
        minimum: 1
        type: integer
      nextCursor:
        example: eyJpZCI6IjEyMyJ9
        type: string
      prevCursor:
        example: eyJpZCI6IjEyMiJ9
        type: string
    type: object
  internal_matching_adapters_http.ListMatchRunsResponse:
    description: Cursor-paginated list of match runs
    properties:
      hasMore:
        example: true
        type: boolean
      items:
        description: List of match runs
        items:
          $ref: '#/definitions/github_com_LerianStudio_matcher_internal_matching_adapters_http_dto.MatchRunResponse'
        maxItems: 200
        type: array
      limit:
        example: 20
        maximum: 200
        minimum: 1
        type: integer
      nextCursor:
        example: eyJpZCI6IjEyMyJ9
        type: string
      prevCursor:
        example: eyJpZCI6IjEyMiJ9
        type: string
    type: object
...
  internal_matching_adapters_http.RunMatchRequest:
    properties:
      mode:
        enum:
        - DRY_RUN
        - COMMIT
        example: DRY_RUN
        type: string
      primarySourceId:
        description: |-
          PrimarySourceID optionally specifies which source is the reference (left) side.
          If provided, directed matching: this source -> left, others -> right.
          If omitted, symmetric matching: sources compared freely (first source used as left internally).
        example: 550e8400-e29b-41d4-a716-446655440000
        format: uuid
        type: string
        x-nullable: true
    required:
    - mode
    type: object
  internal_matching_adapters_http.RunMatchResponse:
    properties:
      runId:
        example: 550e8400-e29b-41d4-a716-446655440000
        format: uuid
        type: string
      status:
        enum:
        - PROCESSING
        - COMPLETED
        - FAILED
        example: PROCESSING
        type: string
    type: object
...
info:
  contact:
    email: support@lerian.studio
    name: Lerian Studio Support
    url: https://discord.gg/DnhqKwkGv3
  description: |-
    Reconciliation engine for the Lerian Studio ecosystem.
    Provides automated transaction matching between Midaz ledger and external systems.
  license:
    name: Lerian Studio General License
    url: https://github.com/LerianStudio/matcher/blob/main/LICENSE.md
  termsOfService: http://swagger.io/terms/
  title: Matcher Reconciliation API
  version: v1.0.0
paths:
...
  /v1/matching/contexts/{contextId}/run:
    post:
      consumes:
      - application/json
      description: Triggers a matching run for a reconciliation context. Supports
        DRY_RUN mode for testing rules without committing results, or COMMIT mode
        for persisting matches.
      operationId: runMatch
      parameters:
      - description: Request ID for tracing
        in: header
        name: X-Request-Id
        type: string
      - description: Context ID
        format: uuid
        in: path
        name: contextId
        required: true
        type: string
      - description: Run match payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_matching_adapters_http.RunMatchRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/internal_matching_adapters_http.RunMatchResponse'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Context not found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Trigger a matching run
      tags:
      - Matching
  /v1/matching/contexts/{contextId}/runs:
    get:
      description: Retrieves a list of match runs for a given reconciliation context,
        sorted by creation time descending.
      operationId: listMatchRuns
      parameters:
      - description: Request ID for tracing
        in: header
        name: X-Request-Id
        type: string
      - description: Context ID
        format: uuid
        in: path
        name: contextId
        required: true
        type: string
      - default: 20
        description: Maximum number of records to return
        in: query
        maximum: 200
        minimum: 1
        name: limit
        type: integer
      - description: Cursor for pagination (opaque)
        in: query
        name: cursor
        type: string
      - description: Sort order
        enum:
        - asc
        - desc
        in: query
        name: sort_order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_matching_adapters_http.ListMatchRunsResponse'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Context not found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List match runs
      tags:
      - Matching
...
  /v1/matching/runs/{runId}:
    get:
      description: Retrieves details of a match run by its ID, including status, timing,
        and aggregate statistics.
      operationId: getMatchRun
      parameters:
      - description: Request ID for tracing
        in: header
        name: X-Request-Id
        type: string
      - description: Run ID
        format: uuid
        in: path
        name: runId
        required: true
        type: string
      - description: Context ID
        format: uuid
        in: query
        name: contextId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_LerianStudio_matcher_internal_matching_adapters_http_dto.MatchRunResponse'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Match run not found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get match run
      tags:
      - Matching
  /v1/matching/runs/{runId}/groups:
    get:
      description: Returns a cursor-paginated list of match groups from a matching
        run, with optional sorting. Each group contains matched transaction pairs
        and confidence scores.
      operationId: getMatchRunGroups
      parameters:
      - description: Request ID for tracing
        in: header
        name: X-Request-Id
        type: string
      - description: Run ID
        format: uuid
        in: path
        name: runId
        required: true
        type: string
      - description: Context ID
        format: uuid
        in: query
        name: contextId
        required: true
        type: string
      - default: 20
        description: Maximum number of records to return
        in: query
        maximum: 200
        minimum: 1
        name: limit
        type: integer
      - description: Cursor for pagination (opaque)
        in: query
        name: cursor
        type: string
      - description: Sort order
        enum:
        - asc
        - desc
        in: query
        name: sort_order
        type: string
      - description: Sort field
        enum:
        - id
        - created_at
        - status
        in: query
        name: sort_by
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_matching_adapters_http.ListMatchGroupsResponse'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Match run not found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get match run results
      tags:
      - Matching
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    description: 'Bearer token authentication (format: "Bearer {token}")'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"